Templates kit: add Duration
